Update yt-dlp-media-ripper.py
Fix 2 bugs.
First bug was with the transcripts.
Second bug was with selecting the options
When you selected number three, it would download the video, but not the audio

Discovered another bug

With an added feature
It would create a subdirectry based on your search query.However, on a failed search, it still creates that subdirectry, which isn't a nonstart.But it is annoying it is something i'm going to work on correcting

Another future edition. I think I'm gonna add some tour. Networking, so it'll change identities like every 15 downloads. That's for security prevent you from getting IP banned.

As well as I'm going to increase the sleep in between downloads to 3 seconds.